+++ Command/UpdateCoreRoutesCommand.php
@@ -49,7 +49,7 @@ class UpdateCoreRoutesCommand extends BaseCommand
 
             $output->writeln(sprintf(" % 5s - % -30s - %s", "ID", "Name", "Url"));
 
-            foreach ($this->getSiteManager()->findBy() as $site) {
+            foreach ($this->getSiteManager()->findBy(array()) as $site) {
                 $output->writeln(sprintf(" % 5s - % -30s - %s", $site->getId(), $site->getName(), $site->getUrl()));
             }