+++ Model/PageManagerInterface.php
@@ -10,27 +10,15 @@
 
 namespace Sonata\PageBundle\Model;
 
+use Sonata\CoreBundle\Entity\ManagerInterface;
+
 /**
  * Defines methods to interact with the persistency layer of a PageInterface
  *
  * @author Thomas Rabaix <thomas.rabaix@sonata-project.org>
  */
-interface PageManagerInterface
+interface PageManagerInterface extends ManagerInterface
 {
-    /**
-     * @param array $criteria
-     *
-     * @return PageInterface
-     */
-    public function findOneBy(array $criteria = array());
-
-    /**
-     * @param array $criteria
-     *
-     * @return array
-     */
-    public function findBy(array $criteria = array());
-
     /**
      * Returns a page with the give slug
      *
@@ -40,18 +28,6 @@ interface PageManagerInterface
      * @return PageInterface
      */
     public function getPageByUrl(SiteInterface $site, $url);
-
-    /**
-     * @param PageInterface $object
-     */
-    public function save(PageInterface $object);
-
-    /**
-     * @param array $params
-     *
-     * @return PageInterface
-     */
-    public function create(array $params = array());
 
     /**
      * Returns an array of Pages Entity where the id is the key
@@ -68,9 +44,4 @@ interface PageManagerInterface
      * @return mixed
      */
     public function fixUrl(PageInterface $page);
-
-    /**
-     * @return string
-     */
-    public function getClass();
 }